+++ lib/src/bindings/bindings_player_web.dart
@@ -550,6 +550,11 @@ class FlutterSoLoudWeb extends FlutterSoLoud {
   }
 
   @override
+  void setInaudibleBehavior(SoundHandle handle, bool mustTick, bool kill) {
+    return wasmSetInaudibleBehavior(handle.id, mustTick ? 1 : 0, kill ? 1 : 0);
+  }
+
+  @override
   int getMaxActiveVoiceCount() {
     return wasmGetMaxActiveVoiceCount();
   }